Implement the parsing of websocket messages into chat messages
For this, I: created a service to parse the message into a paragraph with appropriate styling and databinding, set the message view to use a richtextbox which in the chat would use each message as it's itemtemplate which consists of unfortunately a lot of code behind, combo box messaging and general destiny.gg chat features, a few other things. Significant commit.